backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-decouple==3.8
bcrypt==4.0.1
PyJWT==2.8.0
celery==5.3.4
redis==5.0.1
transformers==4.35.2
torch==2.1.1
openai==1.3.5
sendgrid==6.10.0
twilio==8.10.0
schedule==1.2.0
requests==2.31.0
python-dateutil==2.8.2
